+++ frontend/src/providers/NotificationProvider.tsx
@@ -1,7 +1,7 @@
 'use client'
 
 import React, { createContext, useContext, useState, ReactNode } from 'react'
-import Notification from './Notification'
+import Notification from '../components/Notification'
 
 type NotificationType = 'success' | 'error' | 'info'
